+++ NEWS.md
@@ -1,4 +1,7 @@
-# cauphy
+# cauphy 1.0.3
+
+* Make 'predict.cauphylm' compatible with 'predict.phylolm' version 2.6.5.
+* Fix CRAN check results issues with r-devel.
 
 # cauphy 1.0.2